refactor: move imagelimit functionality in own folder
