+++ ftw/simplelayout/images/configure.zcml
@@ -1,23 +1,8 @@
 <configure
     xmlns="http://namespaces.zope.org/zope"
-    xmlns:zcml="http://namespaces.zope.org/zcml"
-    xmlns:plone="http://namespaces.plone.org/plone"
-    xmlns:browser="http://namespaces.zope.org/browser"
     i18n_domain="ftw.simplelayout">
 
     <include package=".cropping" />
-
-    <adapter for="zope.interface.Interface"
-             factory=".limits.ImageLimits" />
-
-    <adapter for="zope.interface.Interface"
-             factory=".validators.ImageLimitValidatorMessages" />
-
-    <browser:page
-      for="*"
-      name="ftw.simplelayout.images.macros"
-      template="templates/macros.pt"
-      permission="zope2.View"
-      />
+    <include package=".limits" />
 
 </configure>
