+++ ftw/simplelayout/images/limits/browser/configure.zcml
@@ -1,0 +1,14 @@
+<configure
+    xmlns="http://namespaces.zope.org/zope"
+    xmlns:plone="http://namespaces.plone.org/plone"
+    xmlns:browser="http://namespaces.zope.org/browser"
+    i18n_domain="ftw.simplelayout">
+
+    <browser:page
+      for="*"
+      name="ftw.simplelayout.images.macros"
+      template="templates/macros.pt"
+      permission="zope2.View"
+      />
+
+</configure>
